Refactor: issueList 데이터 IssueList에서만 가지고 있도록 수정
- IssueList와 checkReducer에서 가지고 있었음
